+++ app/Http/Controllers/ImageController.php
@@ -11,9 +11,12 @@ use App\Item;
 class ImageController extends Controller
 {
     protected $max_size_to_fit = 800; // width or height in pixels
+    const SECONDS_TO_CACHE = 2592000; // 30days (60sec * 60min * 24hours * 30days)
 
     public function resize($id, $width, $height=null)
     {
+        $headers = ['Cache-Control' => 'max-age=' . self::SECONDS_TO_CACHE];
+
         if (
             ($width <= $this->max_size_to_fit) &&
             ($height <= $this->max_size_to_fit) &&
@@ -30,7 +33,7 @@ class ImageController extends Controller
 
             $imagePath = public_path() . Item::getImagePathForId($id, false, $resize, $resize_method);
 
-            return response()->file($imagePath);
+            return response()->file($imagePath, $headers);
         }
 
         return App::abort(404);