+++ resources/views/katalog.blade.php
@@ -150,6 +150,7 @@
     	                	<a href="{!! $item->getUrl() !!}">
                                 @php
                                     list($width, $height) = getimagesize(public_path() . $item->getImagePath());
+                                    $width =  max($width,1); // prevent division by zero exception
                                 @endphp
                                 <div class="ratio-box" style="padding-bottom: {{ round(($height / $width) * 100, 4) }}%;">
     	                		     @include('components.item_image_responsive', ['item' => $item])